rename Editable{*} to {*}ForEditing
