+++ src/Adapter/CartRule/QueryHandler/GetCartRuleForEditingHandler.php
@@ -37,13 +37,13 @@ use PrestaShop\PrestaShop\Core\Domain\CartRule\Exception\CartRuleException;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\Exception\CartRuleNotFoundException;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\Query\GetCartRuleForEditing;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryHandler\GetCartRuleForEditingHandlerInterface;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRule;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleActions;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleConditions;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleInformation;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleMinimum;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleReduction;
-use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\EditableCartRuleRestrictions;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleActionForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleConditionsForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleInformationForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleMinimumForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleReductionForEditing;
+use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleRestrictionsForEditing;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\ValueObject\DiscountApplicationType;
 use PrestaShop\PrestaShop\Core\Domain\Customer\ValueObject\CustomerId;
 use PrestaShop\PrestaShop\Core\Domain\Customer\ValueObject\NoCustomerId;
@@ -57,12 +57,12 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
     /**
      * @param GetCartRuleForEditing $query
      *
-     * @return EditableCartRule
+     * @return CartRuleForEditing
      *
      * @throws CartRuleException
      * @throws CartRuleNotFoundException
      */
-    public function handle(GetCartRuleForEditing $query): EditableCartRule
+    public function handle(GetCartRuleForEditing $query): CartRuleForEditing
     {
         $cartRuleId = $query->getCartRuleId();
         $cartRule = $this->getCartRule($cartRuleId);
@@ -73,7 +73,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
         $dateAdd = $cartRule->date_add;
         $dateUpd = $cartRule->date_upd;
 
-        return new EditableCartRule(
+        return new CartRuleForEditing(
             $cartRuleId,
             $cartRuleInformation,
             $cartRuleConditions,
@@ -83,9 +83,9 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
         );
     }
 
-    private function getCartRuleInformation(CartRule $cartRule): EditableCartRuleInformation
+    private function getCartRuleInformation(CartRule $cartRule): CartRuleInformationForEditing
     {
-        return new EditableCartRuleInformation(
+        return new CartRuleInformationForEditing(
             $cartRule->name,
             $cartRule->description,
             $cartRule->code,
@@ -96,7 +96,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
         );
     }
 
-    private function getCartRuleConditions(CartRule $cartRule): EditableCartRuleConditions
+    private function getCartRuleConditions(CartRule $cartRule): CartRuleConditionsForEditing
     {
         $customerId = (int) $cartRule->id_customer !== NoCustomerId::NO_CUSTOMER_ID_VALUE ? new CustomerId((int) $cartRule->id_customer) : new NoCustomerId();
         $dateFrom = $cartRule->date_from;
@@ -106,7 +106,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
         if (!empty($cartRule->minimum_amount)) {
             $minimumAmount = new DecimalNumber($cartRule->minimum_amount);
             if (!$minimumAmount->equalsZero()) {
-                $cartRuleMinimum = new EditableCartRuleMinimum(
+                $cartRuleMinimum = new CartRuleMinimumForEditing(
                     $minimumAmount,
                     (bool) $cartRule->minimum_amount_tax,
                     (int) $cartRule->minimum_amount_currency,
@@ -115,7 +115,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
             }
         }
 
-        $cartRuleRestrictions = new EditableCartRuleRestrictions(
+        $cartRuleRestrictions = new CartRuleRestrictionsForEditing(
             (bool) $cartRule->country_restriction,
             (bool) $cartRule->carrier_restriction,
             (bool) $cartRule->group_restriction,
@@ -124,7 +124,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
             (bool) $cartRule->shop_restriction
         );
 
-        return new EditableCartRuleConditions(
+        return new CartRuleConditionsForEditing(
             $customerId,
             !DateTimeUtils::isNull($dateFrom) ? new DateTime($dateFrom) : null,
             !DateTimeUtils::isNull($dateTo) ? new DateTime($dateTo) : null,
@@ -135,7 +135,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
         );
     }
 
-    private function getCartRuleActions(CartRule $cartRule): EditableCartRuleActions
+    private function getCartRuleActions(CartRule $cartRule): CartRuleActionForEditing
     {
         $discountApplicationType = $this->getDiscountApplicationType($cartRule);
 
@@ -144,7 +144,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
             $discountProductId = (int) $cartRule->reduction_product;
         }
 
-        $reduction = new EditableCartRuleReduction(
+        $reduction = new CartRuleReductionForEditing(
             new DecimalNumber($cartRule->reduction_percent),
             new DecimalNumber($cartRule->reduction_amount),
             (bool) $cartRule->reduction_tax,
@@ -153,7 +153,7 @@ final class GetCartRuleForEditingHandler extends AbstractCartRuleHandler impleme
             !$cartRule->reduction_exclude_special
         );
 
-        return new EditableCartRuleActions(
+        return new CartRuleActionForEditing(
             (bool) $cartRule->free_shipping,
             $reduction,
             $this->getDiscountApplicationType($cartRule),
